fastapi-starter/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.21
SQLAlchemy>=2.0
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
passlib[bcrypt,argon2]>=1.7
argon2-cffi>=23.1.0
aiosqlite>=0.20.0
greenlet>=3.0